Added alert for CollectionFS.
I wanted to add this because a lot of people still use CollectionFS and S3 or gridFS to upload files (even though Slingshot seems to picking up a lot of traction).  This is worthwhile because the server will actually crash because of an uncaught exception when trying to manipulate a bad image and/or a file that's no an image.  So this is critically important.

No fix has been submitted so I marked the problem for all versions.